+++ frontend/src/api/meta.ts
@@ -3,6 +3,7 @@ import { apiClient } from "./apiUtils";
 export interface MetaResponse {
   version: string;
   http_server_timeout: string;
+  max_query_limit: number;
 }
 
 export const metaApi = {